# contracts/api.openapi.json
{
  "components": {
    "schemas": {
      "ErrorBody": {
        "properties": {
          "error_code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "error_message": {
            "title": "Error Message",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "error_message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "INVALID_URL",
          "UNSUPPORTED_SITE",
          "LOGIN_REQUIRED",
          "GEO_BLOCKED",
          "FFMPEG_MISSING",
          "NETWORK",
          "BLOCKED_ADDRESS",
          "TOO_LARGE",
          "TOO_LONG",
          "QUEUE_FULL",
          "DISK_FULL",
          "CANCELLED",
          "UNKNOWN"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "ffmpeg": {
            "title": "Ffmpeg",
            "type": "boolean"
          },
          "ffprobe": {
            "title": "Ffprobe",
            "type": "boolean"
          },
          "js_runtime": {
            "anyOf": [
              {
                "enum": [
                  "deno",
                  "node"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Js Runtime"
          },
          "max_concurrent": {
            "title": "Max Concurrent",
            "type": "integer"
          },
          "max_duration_min": {
            "title": "Max Duration Min",
            "type": "integer"
          },
          "max_filesize_mb": {
            "title": "Max Filesize Mb",
            "type": "integer"
          },
          "max_queue": {
            "title": "Max Queue",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "ytdlp_version": {
            "title": "Ytdlp Version",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "version",
          "ytdlp_version",
          "ffmpeg",
          "ffprobe",
          "max_filesize_mb",
          "max_duration_min",
          "max_concurrent",
          "max_queue"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "InfoRequest": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "InfoRequest",
        "type": "object"
      },
      "InfoResponse": {
        "properties": {
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "heights": {
            "items": {
              "type": "integer"
            },
            "title": "Heights",
            "type": "array"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "uploader": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploader"
          }
        },
        "required": [
          "title",
          "heights"
        ],
        "title": "InfoResponse",
        "type": "object"
      },
      "JobCreated": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobCreated",
        "type": "object"
      },
      "JobRequest": {
        "properties": {
          "audio_format": {
            "default": "mp3",
            "enum": [
              "mp3",
              "m4a",
              "opus"
            ],
            "title": "Audio Format",
            "type": "string"
          },
          "height": {
            "default": "best",
            "enum": [
              "360",
              "480",
              "720",
              "1080",
              "1440",
              "2160",
              "best"
            ],
            "title": "Height",
            "type": "string"
          },
          "mode": {
            "enum": [
              "video",
              "audio"
            ],
            "title": "Mode",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "mode"
        ],
        "title": "JobRequest",
        "type": "object"
      },
      "JobStatus": {
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "eta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "percent": {
            "title": "Percent",
            "type": "number"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "status": {
            "enum": [
              "queued",
              "downloading",
              "processing",
              "done",
              "error",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "percent"
        ],
        "title": "JobStatus",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "VideoDownloader",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/api/info": {
      "post": {
        "operationId": "info_api_info_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Info"
      }
    },
    "/api/jobs": {
      "post": {
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too many jobs running or queued"
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not enough free disk space"
          }
        },
        "summary": "Create Job"
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "description": "Cancel a queued or running job. A finished job is discarded together with its file.",
        "operationId": "cancel_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown job"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job"
      },
      "get": {
        "operationId": "job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown job"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Status"
      }
    },
    "/api/jobs/{job_id}/file": {
      "get": {
        "operationId": "job_file_api_jobs__job_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/octet-stream": {}
            },
            "description": "The downloaded file"
          },
          "404": {
            "description": "Unknown job"
          },
          "409": {
            "description": "Job is not finished or failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job File"
      }
    }
  }
}